+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Pact creation fix
+- Organisations to contain associateUsers link
 
 ## [7.0.0] - 2020-03-13